feat: about & tutorial page
