+++ src/app/components/tutorial/tutorial.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-tutorial',
+  standalone: true,
+  imports: [],
+  templateUrl: './tutorial.component.html',
+  styleUrl: './tutorial.component.scss'
+})
+export class TutorialComponent {
+
+}
