+++ src/app/tutorial-page/tutorial-page.component.html
@@ -1,0 +1,13 @@
+<div class="flex flex-col h-full -mb-4">
+    <div class="flex flex-col items-center h-full mt-6 w-full">
+        <div class="flex w-full p-4">
+            <a class="font-bold" href="/">
+                <h6>
+                    <i class="text-xs pi pi-chevron-left"></i>
+                    <span class="ml-4 text-[#224063]">Back to Home page</span>
+                </h6>
+            </a>
+        </div>
+        <app-tutorial></app-tutorial>
+    </div>
+</div>
