+++ src/app/tutorial-page/tutorial-page.component.ts
@@ -1,0 +1,13 @@
+import { Component } from '@angular/core';
+import { TutorialComponent } from "../components/tutorial/tutorial.component";
+
+@Component({
+  selector: 'app-tutorial-page',
+  standalone: true,
+  imports: [TutorialComponent],
+  templateUrl: './tutorial-page.component.html',
+  styleUrl: './tutorial-page.component.scss'
+})
+export class TutorialPageComponent {
+
+}
